Move translate methods into tim_translator.py
